refactor InfoDialog
